More longs

+++ src/Q1/sieve9.c
@@ -53,11 +53,11 @@ int main (int argc, char *argv[])
 
  /* find how many elements are assigned to this process */
  low_value = 2 + BLOCK_LOW(id,p,n-1);
- int proc_low = low_value;
+ unsigned long proc_low = low_value;
  high_value = 2 + BLOCK_HIGH(id,p,n-1);
  size = BLOCK_SIZE(id,p,n-1);
  proc0_size = (n-1)/p;
- sqrtN = (int) sqrt((double) n);
+ sqrtN = (long) sqrt((double) n);
  if ((2 + proc0_size) < sqrtN)
  {
   if (!id)
@@ -107,7 +107,7 @@ int main (int argc, char *argv[])
 
    if (low_value & 1) low_value--;
 
-   int inner_size = high_value - low_value;
+   unsigned long inner_size = high_value - low_value;
    if (inner_size > blocksize) inner_size = blocksize;
 
    do {
